Code cleanup

+++ src/main/java/com/liquid/kochanparser/Time.java
@@ -34,7 +34,7 @@ public class Time
         int result = 0;
         for (int i = 0; i < number.length (); i++)
         {
-            result = result * 10 + (int) (number.charAt (i) - '0');
+            result = result * 10 + number.charAt (i) - '0';
         }
         return result;
     }
